refactor : Lottos 일급컬렉션 사용
