+++ src/main/java/lotto/domain/Lottos.java
@@ -1,0 +1,34 @@
+package lotto.domain;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Lottos {
+    private final List<Lotto> lottos;
+
+    public Lottos(List<Lotto> lottos){
+        this.lottos = lottos;
+    }
+
+    public static Lottos create(MyPurchase myPurchase) {
+        List<Lotto> lottos = new ArrayList<>();
+        for (int i = 0; i < myPurchase.count(); i++) {
+            lottos.add(Lotto.autoGenerate());
+        }
+        return new Lottos(lottos);
+    }
+
+    public int count() {
+        return lottos.size();
+    }
+
+    public Lotto find(int index) {
+        return lottos.get(index);
+    }
+
+    public void checkWin(MyResult myResult, WinLotto winLotto) {
+        for (Lotto lotto : lottos) {
+            myResult.update(lotto.checkMatchingNumbers(winLotto));
+        }
+    }
+}
