+++ src/main/java/lotto/domain/MyLottoGame.java
@@ -4,39 +4,22 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class MyLottoGame {
-    private final List<Lotto> lottos;
+    private final Lottos lottos;
     private final MyResult result;
 
-    public MyLottoGame(List<Lotto> lottos, MyPurchase myPurchase) {
+    public MyLottoGame(Lottos lottos, MyPurchase myPurchase) {
         this.lottos = lottos;
         this.result = new MyResult(myPurchase);
     }
 
     public static MyLottoGame autoGenerate(MyPurchase myPurchase) {
-        List<Lotto> lottos = new ArrayList<>();
-        for (int i = 0; i < myPurchase.count(); i++) {
-            lottos.add(autoGenerateOne());
-        }
+        Lottos lottos = Lottos.create(myPurchase);
         return new MyLottoGame(lottos, myPurchase);
     }
 
-    private static Lotto autoGenerateOne() {
-        return Lotto.autoGenerate();
-    }
-
     public MyResult checkWin(WinLotto winLotto) {
-        for (Lotto lotto : lottos) {
-            result.update(lotto.checkMatchingNumbers(winLotto));
-        }
+        lottos.checkWin(result, winLotto);
         return result;
-    }
-
-    public int count() {
-        return lottos.size();
-    }
-
-    public Lotto find(int index) {
-        return lottos.get(index);
     }
 
     public MyResult result() {
@@ -45,5 +28,9 @@ public class MyLottoGame {
 
     public int countRank(Rank rank) {
         return result().countRank(rank);
+    }
+
+    public Lottos getLottos() {
+        return lottos;
     }
 }
